+++ 2019/src/solutions/day24/part1.rs
@@ -1,6 +1,6 @@
 use super::parse_input;
 
-pub fn solve(input: &str) -> u32 {
+pub fn solve(input: &str) -> usize {
     let input = parse_input(input);
 
     panic!("Not implemented");
